Move tileset manager to separate file
